test(#641): inject a stub Shiprocket transport instead of mocking global.fetch
The CI 401 was the rates route hitting the REAL Shiprocket API with throwaway
test creds: jest.spyOn(global,'fetch') doesn't reliably intercept the in-process
Medusa server's own fetch in CI (it does locally), so the stub was bypassed.

Fix (no global.fetch dependency):
- ShiprocketClient takes an injectable fetchImpl (defaults to a wrapper around
  globalThis.fetch — global.fetch spies in unit tests still work).
- New shiprocket/stub-fetch.ts returns canned auth/pickup/serviceability data.
- resolver injects the stub when SHIPROCKET_STUB=1 (lazy require; inert in prod).
- shiprocket-rates.spec sets SHIPROCKET_STUB=1 and drops the global.fetch mock +
  the un-capturable serviceability-query assertions.

Local: integration 1 passed, 16 related unit tests passed.
